chore: move satellite-animation component
